One handler to check which tab is selected to display in dashboard page
