+++ app/dashboard/[userId]/sidebar-userbar.tsx
@@ -9,11 +9,12 @@ export default function SideBarUserBar() {
   return (
     <div className="pl-6 pr-4 py-4 bg-[#232529] flex items-center justify-between mt-auto">
       <div className="flex items-center">
-        <div className="relative w-8 h-8 rounded-full before:absolute before:w-2 before:h-2 before:bg-green-500 before:rounded-full before:right-0 before:bottom-0 before:ring-1 before:ring-white">
+        <div className="relative w-8 h-8 rounded-full before:absolute before:w-2 before:h-2 before:bg-green-500 before:rounded-full before:right-0 before:bottom-0 before:ring-1 before:ring-white bg-emerald-600">
           <img
             className="rounded-full"
-            src="https://avatars.githubusercontent.com/u/19399457?s=400&u=85e5bfae12dfb6ac040a98178bb35b84ec68d41a&v=4"
-            alt=""
+            // src="https://avatars.githubusercontent.com/u/19399457?s=400&u=85e5bfae12dfb6ac040a98178bb35b84ec68d41a&v=4"
+            src={`https://api.dicebear.com/5.x/open-peeps/svg?seed=${session?.user?.name}&flip=true&translateX=5`}
+            alt="avatar"
           />
         </div>
         <div className="flex flex-col pl-3">
